// libs/frontend/src/cms-plugin-types.ts
export interface Media {
  id: string;
  category?: (string | null) | MediaCategory;
  comment?: string | null;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Banner {
  id: string;
  message: string;
  cta?: {
    show?: boolean | null;
    label?: string | null;
    link?: NewLink;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "NewLink".
 */
export interface NewLink {
  linkType: 'custom' | 'internal';
  doc?: (string | null) | Page;
  queryString?: string | null;
  fragment?: string | null;
  url?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: string;
  /**
   * A hero section is the first thing a user sees when they visit a page. Only one hero block can be added to a page. To replace the current hero block by a different block type, remove it and add a new one.
   */
  hero?: (HeroSlides | HeroVideo | HeroHeading)[] | null;
  /**
   * Add blocks to fill the page with content. You can reorder the blocks by dragging and dropping them using the handle on the left side.
   */
  content?:
    | (
        | LeadText
        | ImageWithFloatingText
        | Story
        | Features
        | Separator
        | WideImage
        | TextColumnsWithImages
        | Map
        | Testimonials
      )[]
    | null;
  seo?: {
    /**
     * The description is shown in search engine results. It should be between 100 and 150 characters.
     */
    description?: string | null;
    /**
     * The image is shown in search engine results and when the page is shared on social media. It will be automatically sized to 1200x630 pixels.
     */
    image?: (string | null) | Media;
  };
  /**
   * Choose the brand to which the page belongs. The brand determines the theme of the page.
   */
  brand: string | Brand;
  /**
   * The pathname is used to navigate to this page. It must be unique. The first path segment must be the brand's home link.
   */
  pathname: string;
  pathname_locked?: boolean | null;
  pathname_createRedirect?: boolean | null;
  /**
   * The title is shown in the title bar of the browser and in search engine results. Include important keywords for SEO. The brand’s base title is appended to the title.
   */
  title?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroSlides".
 */
export interface HeroSlides {
  /**
   * This heading is only used for SEO purposes and is not shown on the page. Since the slide overlay titles semantically don't define the page's main heading, you can use this field to define the main heading of the page.
   */
  seoPageHeading: string;
  /**
   * You can use this either for a single hero image or a slideshow with multiple images. Each slide can have an overlay title and a call-to-action button. You can add up to six slides.
   */
  slides: {
    image: string | Media;
    /**
     * Depending on the user’s device or window size, the slide often has a different aspect ratio than the image. This setting defines how the image is aligned within the slide area.
     */
    imageAlignment?: ('center' | 'bottom') | null;
    overlayTitle?: {
      show?: boolean | null;
      text?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      supportingText?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      cta?: {
        show?: boolean | null;
        label?: string | null;
        link?: NewLink;
        variant?: ('primary' | 'secondary') | null;
      };
      position?: ('center' | 'top-left' | 'top-right' | 'bottom-right' | 'bottom-left') | null;
      /**
       * The overlay is a semi-transparent black layer that is placed on top of the image to make the text more readable. Choose the intensity that is the best trade-off between readability of the text and brightness of the image.
       */
      overlay?: ('subtle' | 'moderate' | 'intense') | null;
    };
    id?: string | null;
  }[];
  /**
   * Depending on the amount of information in the slides (title, CTA), different intervals might be more suitable. The default value is 7 seconds.
   */
  autoplayIntervalInSeconds?: number | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'HeroSlides';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroVideo".
 */
export interface HeroVideo {
  /**
   * The video should be optimized for web pages before uploading it.
   */
  video: string | Media;
  /**
   * The preview image is shown while the video is still loading. It should be the first frame of the video to provide a seamless transition.
   */
  previewImage?: (string | null) | Media;
  overlayTitle?: {
    show?: boolean | null;
    text?: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    supportingText?: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    cta?: {
      show?: boolean | null;
      label?: string | null;
      link?: NewLink;
      variant?: ('primary' | 'secondary') | null;
    };
    position?: ('center' | 'top-left' | 'top-right' | 'bottom-right' | 'bottom-left') | null;
    /**
     * The overlay is a semi-transparent black layer that is placed on top of the image to make the text more readable. Choose the intensity that is the best trade-off between readability of the text and brightness of the image.
     */
    overlay?: ('subtle' | 'moderate' | 'intense') | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'HeroVideo';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroHeading".
 */
export interface HeroHeading {
  heading: string;
  image?: (string | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'HeroHeading';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "LeadText".
 */
export interface LeadText {
  heading?: string | null;
  text: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  cta?: {
    show?: boolean | null;
    label?: string | null;
    link?: NewLink;
    variant?: ('primary' | 'secondary') | null;
  };
  elementId?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'LeadText';
}
// ...
export interface Separator {
  id?: string | null;
  blockName?: string | null;
  blockType: 'Separator';
}
// ...
export interface Brand {
  id: string;
  name: string;
  homeLink?: NewLink;
  /**
   * The base title is appended to the titles of the brand’s pages. If the page does not have a title, the base title will be used as the title. Include important keywords in the title for SEO.
   */
  baseTitle?: string | null;
  logo: string | Media;
  /**
   * A banner is useful to announce promotions or important news and can have a call to action. It will be shown on all pages of the brand.
   */
  banner?: (string | null) | Banner;
  navLinks?:
    | {
        label: string;
        link: NewLink;
        id?: string | null;
      }[]
    | null;
  bookCta?: {
    show?: boolean | null;
    label?: string | null;
    link?: NewLink;
  };
  footer?: {
    linkGroups?:
      | {
          title: string;
          links?:
            | {
                label: string;
                link: NewLink;
                id?: string | null;
              }[]
            | null;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "redirects".
 */
export interface Redirect {
  id: string;
  fromPathname: string;
  to: {
    page: string | Page;
    queryString?: string | null;
    fragment?: string | null;
  };
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Settings {
  id: string;
  publishedLocales: {
    /**
     * Select locales with completely translated content to make them available to the user on the website. To add a new locale, it must be first added to the 'Locale Configurations' collection.
     */
    publishedLocales: (string | LocaleConfig)[];
    /**
     * The fallback locale is used when a translation is not available in the requested locale. It must be one of the published locales.
     */
    fallbackLocale: string | LocaleConfig;
  };
  /**
   * Hide the complete website and show a maintenance screen instead.
   */
  maintenanceScreen?: {
    show?: boolean | null;
    message?: string | null;
  };
  maps?: {
    /**
     * Enter the region code for maps, e.g. CO for Colombia. Must be two letters in uppercase. See https://developers.google.com/maps/coverage#coverage-legend
     */
    region?: string | null;
    /**
     * Enter the ID of the map to display. This is the ID of the map in the Google Maps Platform and defines styling and POI settings.
     */
    mapId?: string | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "common".
 */
export interface Common {
  id: string;
  pageNotFoundScreen?: {
    heading?: string | null;
    text?: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
  };
  errorScreen?: {
    heading?: string | null;
    text?: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
  };
  uiLabels?: {
    banner?: {
      dismiss?: string | null;
    };
    slidesBlock?: {
      /**
       * Use the {{slide}} placeholder to display the slide number.
       */
      goToSlide?: string | null;
    };
    errorBoundary?: {
      title?: string | null;
      text?: string | null;
    };
    maintenanceScreen?: {
      login?: string | null;
    };
    login?: {
      email?: string | null;
      password?: string | null;
      submit?: string | null;
    };
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}